Added custom color logger class to util.py
