+++ tardis/__init__.py
@@ -1,6 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import logging
-import color
+from tardis.io.util import ColoredLogger
 # Affiliated packages may add whatever they like to this file, but
 # should keep this content at the top.
 # ----------------------------------------------------------------------------
@@ -12,7 +12,7 @@ from tardis.io.util import yaml_load_config_file as yaml_load
 
 logging.captureWarnings(True)
 logger = logging.getLogger('tardis')
-logging.setLoggerClass(color.ColoredLogger)
+logging.setLoggerClass(ColoredLogger)
 logger.setLevel(logging.INFO)
 console_handler = logging.StreamHandler()
 console_formatter = logging.Formatter('%(name)s - %(levelname)s - %(message)s')
