+++ tardis/io/util.py
@@ -221,3 +221,63 @@ def to_hdf(path_or_buf, path, elements, complevel=9, complib='blosc'):
             if scalars_path in store:
                 scalars_series = store[scalars_path].append(scalars_series)
         scalars_series.to_hdf(path_or_buf, os.path.join(path, 'scalars'))
+
+
+def formatter_message(message, use_color=True):
+
+    #These are the sequences need to get colored ouput
+    RESET_SEQ = "\033[0m"
+    BOLD_SEQ = "\033[1m"
+    if use_color:
+        message = message.replace(
+            "$RESET", RESET_SEQ).replace("$BOLD", BOLD_SEQ)
+    else:
+        message = message.replace("$RESET", "").replace("$BOLD", "")
+    return message
+
+
+class ColoredFormatter(logging.Formatter):
+    '''
+    Custom logger class for changing levels color
+    '''
+
+    def __init__(self, msg, use_color=True):
+        logging.Formatter.__init__(self, msg)
+        self.use_color = use_color
+
+    def format(self, record):
+        COLOR_SEQ = "\033[1;%dm"
+        RESET_SEQ = "\033[0m"
+        BLACK, RED, GREEN, YELLOW, BLUE, MAGENTA, CYAN, WHITE = range(8)
+        COLORS = {
+            'WARNING': YELLOW,
+            'INFO': WHITE,
+            'DEBUG': BLUE,
+            'CRITICAL': YELLOW,
+            'ERROR': RED
+        }
+        levelname = record.levelname
+        if self.use_color and levelname in COLORS:
+            levelname_color = COLOR_SEQ % (
+                30 + COLORS[levelname]) + levelname + RESET_SEQ
+            record.levelname = levelname_color
+        return logging.Formatter.format(self, record)
+
+
+class ColoredLogger(logging.Logger):
+    '''
+    Custom logger class with multiple destinations
+    '''
+    FORMAT = "[$BOLD%(name)-20s$RESET][%(levelname)-18s]  %(message)s ($BOLD%(filename)s$RESET:%(lineno)d)"
+    COLOR_FORMAT = formatter_message(FORMAT, True)
+
+    def __init__(self, name):
+        logging.Logger.__init__(self, name, logging.DEBUG)
+
+        color_formatter = ColoredFormatter(self.COLOR_FORMAT)
+
+        console = logging.StreamHandler()
+        console.setFormatter(color_formatter)
+
+        self.addHandler(console)
+        return
